Made compatible for both windows and *nix

+++ project.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 from __future__ import print_function
+import platform
 import traceback
 import errno
 import filecmp
@@ -26,7 +27,7 @@ import sys
 import tarfile
 import tempfile
 import time
-import win32file
+if platform.system()=="Windows": import win32file
 
 from color import Coloring
 from git_command import GitCommand, git_require
@@ -53,10 +54,11 @@ def _lwrite(path, content):
   finally:
     fd.close()
 
-  try:
-    os.remove(path)
-  except:
-    pass
+  if platform.system()=="Windows":
+    try:
+      os.remove(path)
+    except:
+      pass
 
   try:
     os.rename(lock, path)
@@ -2066,7 +2068,10 @@ class Project(object):
           _error("%s: Not replacing %s hook", self.relpath, name)
           continue
       try:
-        win32file.CreateSymbolicLink(dst, stock_hook, 1 if os.path.isdir(stock_hook) else 0)
+        if platform.system()=="Windows":
+          win32file.CreateSymbolicLink(dst, stock_hook, 1 if os.path.isdir(stock_hook) else 0)
+        else:
+          os.symlink(os.path.relpath(stock_hook, os.path.dirname(dst)), dst)
       except OSError as e:
         if e.errno == errno.EPERM:
           raise GitError('filesystem must support symlinks')
@@ -2145,7 +2150,10 @@ class Project(object):
           os.makedirs(src)
 
         if name in to_symlink:
-          win32file.CreateSymbolicLink(dst, src, 1 if os.path.isdir(src) else 0)
+          if platform.system()=="Windows":
+            win32file.CreateSymbolicLink(dst, src, 1 if os.path.isdir(src) else 0)
+          else:
+            os.symlink(os.path.relpath(src, os.path.dirname(dst)), dst)
         elif copy_all and not os.path.islink(dst):
           if os.path.isdir(src):
             shutil.copytree(src, dst)
